+++ mdObject/src/test/classes/EmrContent.spec.ts
@@ -1,4 +1,5 @@
 import { EmrContent } from '../../classes';
+import { ObjectState } from '../../enums';
 import { MockEmrMel, mockResultEmr, MockWindow } from '../mocks/mocks';
 
 describe('Class: EmrContent', () => {
@@ -7,85 +8,84 @@ describe('Class: EmrContent', () => {
     let mockEmrMel = new MockEmrMel();
     let _window = new MockWindow();
     let result: string = mockResultEmr;
-    let value: string = '1';
 
     describe('right value', () => {
         beforeAll(() => {
-            component = new EmrContent(result, mockEmrMel as any, _window);
+            component = new EmrContent(result, mockEmrMel as any);
         })
 
         it('check contentId', () => {
             expect(component.contentId).toEqual('0.0');
         });
-        it('check key', () => {
-            expect(component.key).toEqual('0.1');
+        it('check namespace', () => {
+            expect(component.namespace).toEqual('0.1');
         });
-        it('check name', () => {
-            expect(component.name).toEqual('0.2');
+        it('check nodeName', () => {
+            expect(component.nodeName).toEqual('0.2');
         });
-        it('check value', () => {
-            expect(component.value).toEqual('0.3');
+        it('check displayName', () => {
+            expect(component.displayName).toEqual('0.3');
         });
-        it('check _unk0', () => {
-            expect(component._unk0).toEqual('0.4');
+        it('check sourceName', () => {
+            expect(component.sourceName).toEqual('0.4');
         });
-        it('check _unk1', () => {
-            expect(component._unk1).toEqual('0.5');
+        it('check code', () => {
+            expect(component.code).toEqual('0.5');
         });
-        it('check _unk2', () => {
-            expect(component._unk2).toEqual('0.6');
+        it('check codeType', () => {
+            expect(component.codeType).toEqual('0.6');
         });
-        it('check _unk3', () => {
-            expect(component._unk3).toEqual('0.7');
+        it('check contentGroup', () => {
+            expect(component.contentGroup).toEqual('0.7');
         });
-        it('check _unk4', () => {
-            expect(component._unk4).toEqual('0.8');
+        it('check listOrder', () => {
+            expect(component.listOrder).toEqual('0.8');
         });
-        it('check _unk5', () => {
-            expect(component._unk5).toEqual('0.9');
+        it('check contentDefault', () => {
+            expect(component.contentDefault).toEqual('0.9');
         });
     })
 
     describe('default value', () => {
         beforeAll(() => {
-            component = new EmrContent('', mockEmrMel as any, _window);
+            component = new EmrContent('', mockEmrMel as any);
         })
 
         it('check contentId', () => {
             expect(component.contentId).toEqual('');
         });
-        it('check key', () => {
-            expect(component.key).toEqual('');
+        it('check namespace', () => {
+            expect(component.namespace).toEqual('');
         });
-        it('check name', () => {
-            expect(component.name).toEqual('');
+        it('check nodeName', () => {
+            expect(component.nodeName).toEqual('');
         });
-        it('check value', () => {
-            expect(component.value).toEqual('');
+        it('check displayName', () => {
+            expect(component.displayName).toEqual('');
         });
-        it('check _unk0', () => {
-            expect(component._unk0).toEqual('');
+        it('check sourceName', () => {
+            expect(component.sourceName).toEqual('');
         });
-        it('check _unk1', () => {
-            expect(component._unk1).toEqual('');
+        it('check code', () => {
+            expect(component.code).toEqual('');
         });
-        it('check _unk2', () => {
-            expect(component._unk2).toEqual('');
+        it('check codeType', () => {
+            expect(component.codeType).toEqual('');
         });
-        it('check _unk3', () => {
-            expect(component._unk3).toEqual('');
+        it('check contentGroup', () => {
+            expect(component.contentGroup).toEqual('');
         });
-        it('check _unk4', () => {
-            expect(component._unk4).toEqual('');
+        it('check listOrder', () => {
+            expect(component.listOrder).toEqual('');
         });
-        it('check _unk5', () => {
-            expect(component._unk5).toEqual('');
+        it('check contentDefault', () => {
+            expect(component.contentDefault).toEqual('');
         });
     })
 
     describe('methods', () => {
         beforeEach(() => {
-            component = new EmrContent(result, mockEmrMel as any, _window);
+            component = new EmrContent(result, mockEmrMel as any);
             spyOn(component._mel, 'melFunc')
                 .and
                 .returnValue('');
@@ -93,36 +93,32 @@ describe('Class: EmrContent', () => {
 
         describe('toAddString', () => {
             it('right value', () => {
-                spyOn(_window, 'btoa')
-                    .and
-                    .returnValue('123');
-                let _result = component.toAddString(value);
-                let __result = component.key + '^' +
-                    component.name + '^' +
-                    '123' + '^' +
-                    component._unk0 + '^' +
-                    component._unk1 + '^' +
-                    component._unk2 + '^' +
-                    component._unk3 + '^' +
-                    component._unk4 + '^' +
-                    component._unk5;
+                let _result = component.toAddString();
+                let __result = component.namespace + '^' +
+                    component.nodeName + '^' +
+                    component.displayName + '^' +
+                    component.sourceName + '^' +
+                    component.code + '^' +
+                    component.codeType + '^' +
+                    component.contentGroup + '^' +
+                    component.listOrder + '^' +
+                    component.contentDefault;
                 expect(_result).toEqual(__result);
-                expect(_window.btoa).toHaveBeenCalledWith(component.value);
             })
             it('wrong value', () => {
                 spyOn(_window, 'btoa')
                     .and
                     .returnValue('');
-                let _result = component.toAddString(null);
-                let __result = component.key + '^' +
-                    component.name + '^' +
-                    component.value + '^' +
-                    component._unk0 + '^' +
-                    component._unk1 + '^' +
-                    component._unk2 + '^' +
-                    component._unk3 + '^' +
-                    component._unk4 + '^' +
-                    component._unk5;
+                let _result = component.toAddString();
+                let __result = component.namespace + '^' +
+                    component.nodeName + '^' +
+                    component.displayName + '^' +
+                    component.sourceName + '^' +
+                    component.code + '^' +
+                    component.codeType + '^' +
+                    component.contentGroup + '^' +
+                    component.listOrder + '^' +
+                    component.contentDefault;
                 expect(_result).toEqual(__result);
                 expect(_window.btoa).not.toHaveBeenCalled();
             })
@@ -131,28 +127,30 @@ describe('Class: EmrContent', () => {
             beforeEach(() => {
                 spyOn(component, 'toAddString')
                     .and
-                    .returnValue(component.toAddString(value));
+                    .returnValue(component.toAddString());
             })
 
-            it('isNew == true', () => {
-                component = new EmrContent(null, mockEmrMel as any, _window);
+            it('ObjectState.Add', () => {
+                component = new EmrContent(null, mockEmrMel as any);
+                component.state = ObjectState.Add;
+
                 spyOn(component, 'toAddString')
                     .and
-                    .returnValue(component.toAddString(value));
-                component.save(value);
-                expect(component._mel.melFunc).toHaveBeenCalledWith('{MEL_ADD_CONTENT("' + component.toAddString(value) + '")}');
-                expect((component as any).isNew).toEqual(false);
+                    .returnValue(component.toAddString());
+                component.save();
+                expect(component._mel.melFunc).toHaveBeenCalledWith('{MEL_ADD_CONTENT("' + component.toAddString() + '")}');
                 expect(component.toAddString).toHaveBeenCalled();
             })
-            it('isNew == false', () => {
-                component.save(value);
-                expect(component._mel.melFunc).toHaveBeenCalledWith('{MEL_REMOVE_CONTENT("' + component.contentId + '")}');
-                expect(component._mel.melFunc).toHaveBeenCalledWith('{MEL_ADD_CONTENT("' + component.toAddString(value) + '")}');
+            it('ObjectState.Update', () => {
+                component.state = ObjectState.Update;
+                component.save();
+                expect(component._mel.melFunc).toHaveBeenCalledWith('{MEL_UPDATE_CONTENT("' + component.contentId + '","' + component.toChangeString() +  '")}');
                 expect(component.toAddString).toHaveBeenCalled();
             })
         })
         it('remove', () => {
-            component.remove();
+            component.state = ObjectState.Remove;
+            component.save();
             expect(component._mel.melFunc).toHaveBeenCalledWith('{MEL_REMOVE_CONTENT("' + component.contentId + '")}');
         })
     })